update package name
