--- mongocryptd.pid
@@ -1,1 +1,0 @@
-{ "port" : 27020, "pid" : 6498 }
