+++ src/main/java/com/example/demofle/mongoTemplate/Customer.java
@@ -1,4 +1,4 @@
-package com.example.demofle.templ;
+package com.example.demofle.mongoTemplate;
 
 import org.springframework.data.annotation.Id;
