+++ src/main/java/com/example/demofle/repository/AccessingDataMongodbApplication.java
@@ -1,4 +1,4 @@
-package com.example.demofle.repo;
+package com.example.demofle.repository;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.CommandLineRunner;
