+++ src/main/resources/application.properties
@@ -1,3 +1,1 @@
-#spring.data.mongodb.uri=mongodb+srv://caspar:c@bil.cxn3q.mongodb.net/myFirstDatabase?retryWrites=true&w=majority
-
 spring.data.mongodb.uri=mongodb://c:c@13.214.135.136:27077/test?authSource=admin
