--- src/test/java/com/example/demofle/demofle/DemofleApplicationTests.java
@@ -1,13 +1,0 @@
-package com.example.demofle.demofle;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class DemofleApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
